refactor: 포지션과 관련된 클래스를 패키지화
