+++ src/main/java/penaltykick/domain/position/KeeperPositions.java
@@ -1,8 +1,6 @@
-package penaltykick.domain;
+package penaltykick.domain.position;
 
 import java.util.List;
-import penaltykick.domain.position.PenaltyPositions;
-import penaltykick.domain.position.PostPosition;
 
 public final class KeeperPositions {
     private final PenaltyPositions keeperPositions;
