+++ src/main/java/penaltykick/domain/position/PostPosition.java
@@ -1,4 +1,4 @@
-package penaltykick.domain;
+package penaltykick.domain.position;
 
 import java.util.Arrays;
